Unify Backend URL
It used to be that each team had their own port number for the backend.
This won't work in production, so there's now an APIURL global.
This global must be imported and included in the request URL of every request being made in the fashion of `${APIURL}/path/to/resource`.
It takes care of the whole 'http://localhost:[0-9][0-9][0-9][0-9]' thing that was going on previously.